requirements.txt
annotated-types==0.5.0
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.1
cymem==2.0.7
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.24.4
packaging==23.1
pathy==0.10.2
preshed==3.0.8
pydantic==2.1.1
pydantic-core==2.4.0
pyspellchecker==0.7.2
python-dotenv==1.0.0
requests==2.31.0
smart-open==6.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.12
tqdm==4.66.1
typer==0.9.0
typing-extensions==4.7.1
urllib3==2.0.4
wasabi==1.1.2
